requirements.txt
itsdangerous==1.1.0
Django==2.0.4
djangorestframework==3.10.3
